pyproject.toml
[tool.poetry]
name = "drf-django"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
config = "^0.5.1"
django = "^5.0.4"
loadenv = "^0.1.1"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"
ipython = "^8.24.0"
python-decouple = "^3.8"
flake8 = "^7.0.0"
django-bootstrap4 = "^24.3"
django-redis = "^5.4.0"
pillow = "^10.3.0"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.2"
drf-yasg = {extras = ["validation"], version = "^1.21.7"}
drf-spectacular = {extras = ["sidecar"], version = "^0.27.2"}
pytest = "^8.2.1"
pytest-django = "^4.8.0"
requests-mock = "^1.12.1"
stripe = "^9.7.0"
pytest-mock = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
